Move lingering types
